Extract common logic into separate service in Wallet module
