+++ UnstoppableWallet/UnstoppableWallet/Modules/Wallet/WalletCommonService.swift
@@ -1,0 +1,154 @@
+import Foundation
+import RxSwift
+import RxRelay
+import HsToolKit
+import StorageKit
+
+class WalletCommonService {
+    private let keySortType = "wallet-sort-type"
+
+    private let accountManager: AccountManager
+    private let accountRestoreWarningManager: AccountRestoreWarningManager
+    private let reachabilityManager: IReachabilityManager
+    private let balancePrimaryValueManager: BalancePrimaryValueManager
+    private let balanceHiddenManager: BalanceHiddenManager
+    private let cloudAccountBackupManager: CloudAccountBackupManager
+    private let rateAppManager: RateAppManager
+    private let localStorage: StorageKit.ILocalStorage
+    private let sorter = WalletSorter()
+    private let disposeBag = DisposeBag()
+
+    private let activeAccountRelay = PublishRelay<Account?>()
+    private let accountsLostRelay = PublishRelay<()>()
+
+    private let sortTypeRelay = PublishRelay<WalletModule.SortType>()
+    var sortType: WalletModule.SortType {
+        didSet {
+            sortTypeRelay.accept(sortType)
+            localStorage.set(value: sortType.rawValue, for: keySortType)
+        }
+    }
+
+    init(accountManager: AccountManager, accountRestoreWarningManager: AccountRestoreWarningManager, reachabilityManager: IReachabilityManager, balancePrimaryValueManager: BalancePrimaryValueManager, balanceHiddenManager: BalanceHiddenManager, cloudAccountBackupManager: CloudAccountBackupManager, rateAppManager: RateAppManager, localStorage: StorageKit.ILocalStorage) {
+        self.accountManager = accountManager
+        self.accountRestoreWarningManager = accountRestoreWarningManager
+        self.reachabilityManager = reachabilityManager
+        self.balancePrimaryValueManager = balancePrimaryValueManager
+        self.balanceHiddenManager = balanceHiddenManager
+        self.cloudAccountBackupManager = cloudAccountBackupManager
+        self.rateAppManager = rateAppManager
+        self.localStorage = localStorage
+
+        if let rawValue: String = localStorage.value(for: keySortType), let sortType = WalletModule.SortType(rawValue: rawValue) {
+            self.sortType = sortType
+        } else if let rawValue: Int = localStorage.value(for: "balance_sort_key"), rawValue < WalletModule.SortType.allCases.count {
+            // todo: temp solution for restoring from version 0.22
+            sortType = WalletModule.SortType.allCases[rawValue]
+        } else {
+            sortType = .balance
+        }
+
+        subscribe(disposeBag, accountManager.activeAccountObservable) { [weak self] in
+            self?.activeAccountRelay.accept($0)
+        }
+        subscribe(disposeBag, accountManager.accountUpdatedObservable) { [weak self] in
+            self?.handleUpdated(account: $0)
+        }
+        subscribe(disposeBag, accountManager.accountDeletedObservable) { [weak self] in
+            self?.handleDeleted(account: $0)
+        }
+        subscribe(disposeBag, accountManager.accountsLostObservable) { [weak self] isAccountsLost in
+            if isAccountsLost {
+                self?.accountsLostRelay.accept(())
+            }
+        }
+    }
+
+    private func handleUpdated(account: Account) {
+        if account.id == accountManager.activeAccount?.id {
+            activeAccountRelay.accept(account)
+        }
+    }
+
+    private func handleDeleted(account: Account) {
+        accountRestoreWarningManager.removeIgnoreWarning(account: account)
+    }
+
+}
+
+extension WalletCommonService {
+
+    var activeAccountObservable: Observable<Account?> {
+        activeAccountRelay.asObservable()
+    }
+
+    var accountsLostObservable: Observable<()> {
+        accountsLostRelay.asObservable()
+    }
+
+    var sortTypeObservable: Observable<WalletModule.SortType> {
+        sortTypeRelay.asObservable()
+    }
+
+    var balancePrimaryValueObservable: Observable<BalancePrimaryValue> {
+        balancePrimaryValueManager.balancePrimaryValueObservable
+    }
+
+    var balancePrimaryValue: BalancePrimaryValue {
+        balancePrimaryValueManager.balancePrimaryValue
+    }
+
+    var balanceHiddenObservable: Observable<Bool> {
+        balanceHiddenManager.balanceHiddenObservable
+    }
+
+    var activeAccount: Account? {
+        accountManager.activeAccount
+    }
+
+    var watchAccount: Bool {
+        accountManager.activeAccount?.watchAccount ?? false
+    }
+
+    var lastCreatedAccount: Account? {
+        accountManager.popLastCreatedAccount()
+    }
+
+    var balanceHidden: Bool {
+        balanceHiddenManager.balanceHidden
+    }
+
+    var isReachable: Bool {
+        reachabilityManager.isReachable
+    }
+
+    func sort<T: IBalanceItem>(balanceItems: [T]) -> [T] {
+        sorter.sort(balanceItems: balanceItems, sortType: sortType)
+    }
+
+    func toggleBalanceHidden() {
+        balanceHiddenManager.toggleBalanceHidden()
+    }
+
+    func notifyAppear() {
+        rateAppManager.onBalancePageAppear()
+    }
+
+    func notifyDisappear() {
+        rateAppManager.onBalancePageDisappear()
+    }
+
+    func isCloudBackedUp(account: Account) -> Bool {
+        cloudAccountBackupManager.backedUp(uniqueId: account.type.uniqueId())
+    }
+
+    func didIgnoreAccountWarning() {
+        guard let account = accountManager.activeAccount, account.nonRecommended else {
+            return
+        }
+
+        accountRestoreWarningManager.setIgnoreWarning(account: account)
+        activeAccountRelay.accept(account)
+    }
+
+}
