+++ UnstoppableWallet/UnstoppableWallet/Modules/Wallet/WalletService.swift
@@ -3,34 +3,20 @@ import RxSwift
 import RxRelay
 import MarketKit
 import CurrencyKit
-import StorageKit
 import EvmKit
 import HsToolKit
 
 class WalletService {
-    private let keySortType = "wallet-sort-type"
-
+    private let commonService: WalletCommonService
     private let adapterService: WalletAdapterService
     private let coinPriceService: WalletCoinPriceService
     private let cacheManager: EnabledWalletCacheManager
-    private let accountManager: AccountManager
-    private let accountRestoreWarningManager: AccountRestoreWarningManager
-    private let cloudAccountBackupManager: CloudAccountBackupManager
     private let walletManager: WalletManager
     private let marketKit: MarketKit.Kit
-    private let localStorage: StorageKit.ILocalStorage
-    private let rateAppManager: RateAppManager
-    private let balancePrimaryValueManager: BalancePrimaryValueManager
-    private let balanceHiddenManager: BalanceHiddenManager
     private let balanceConversionManager: BalanceConversionManager
     private let feeCoinProvider: FeeCoinProvider
-    private let reachabilityManager: IReachabilityManager
-    private let sorter = WalletSorter()
     private let disposeBag = DisposeBag()
     private var walletDisposeBag = DisposeBag()
-
-    private let activeAccountRelay = PublishRelay<Account?>()
-    private let accountsLostRelay = PublishRelay<()>()
 
     private let totalItemRelay = PublishRelay<WalletModule.TotalItem?>()
     private(set) var totalItem: WalletModule.TotalItem? {
@@ -50,7 +36,7 @@ class WalletService {
 
     private var items: [Item] = [] {
         didSet {
-            let hideZeroBalances = accountManager.activeAccount?.type.hideZeroBalances ?? false
+            let hideZeroBalances = commonService.activeAccount?.type.hideZeroBalances ?? false
 
             if hideZeroBalances {
                 balanceItems = items.filter { $0.balanceData.balanceTotal != 0 || $0.wallet.token.type == .native }
@@ -60,56 +46,20 @@ class WalletService {
         }
     }
 
-    private let sortTypeRelay = PublishRelay<WalletModule.SortType>()
-    var sortType: WalletModule.SortType {
-        didSet {
-            sortTypeRelay.accept(sortType)
-            handleUpdateSortType()
-            localStorage.set(value: sortType.rawValue, for: keySortType)
-        }
-    }
-
     private let queue = DispatchQueue(label: "io.horizontalsystems.unstoppable.wallet-service", qos: .userInitiated)
 
-    init(adapterService: WalletAdapterService, coinPriceService: WalletCoinPriceService, cacheManager: EnabledWalletCacheManager, accountManager: AccountManager, accountRestoreWarningManager: AccountRestoreWarningManager, cloudAccountBackupManager: CloudAccountBackupManager, walletManager: WalletManager, marketKit: MarketKit.Kit, localStorage: StorageKit.ILocalStorage, rateAppManager: RateAppManager, balancePrimaryValueManager: BalancePrimaryValueManager, balanceHiddenManager: BalanceHiddenManager, balanceConversionManager: BalanceConversionManager, appManager: IAppManager, feeCoinProvider: FeeCoinProvider, reachabilityManager: IReachabilityManager) {
+    init(commonService: WalletCommonService, adapterService: WalletAdapterService, coinPriceService: WalletCoinPriceService, cacheManager: EnabledWalletCacheManager, walletManager: WalletManager, marketKit: MarketKit.Kit, balanceConversionManager: BalanceConversionManager, appManager: IAppManager, feeCoinProvider: FeeCoinProvider) {
+        self.commonService = commonService
         self.adapterService = adapterService
         self.coinPriceService = coinPriceService
         self.cacheManager = cacheManager
-        self.accountManager = accountManager
-        self.accountRestoreWarningManager = accountRestoreWarningManager
-        self.cloudAccountBackupManager = cloudAccountBackupManager
         self.walletManager = walletManager
         self.marketKit = marketKit
-        self.localStorage = localStorage
-        self.rateAppManager = rateAppManager
-        self.balancePrimaryValueManager = balancePrimaryValueManager
-        self.balanceHiddenManager = balanceHiddenManager
         self.balanceConversionManager = balanceConversionManager
         self.feeCoinProvider = feeCoinProvider
-        self.reachabilityManager = reachabilityManager
 
-        if let rawValue: String = localStorage.value(for: keySortType), let sortType = WalletModule.SortType(rawValue: rawValue) {
-            self.sortType = sortType
-        } else if let rawValue: Int = localStorage.value(for: "balance_sort_key"), rawValue < WalletModule.SortType.allCases.count {
-            // todo: temp solution for restoring from version 0.22
-            sortType = WalletModule.SortType.allCases[rawValue]
-        } else {
-            sortType = .balance
-        }
-
-        subscribe(disposeBag, accountManager.activeAccountObservable) { [weak self] in
-            self?.activeAccountRelay.accept($0)
-        }
-        subscribe(disposeBag, accountManager.accountUpdatedObservable) { [weak self] in
-            self?.handleUpdated(account: $0)
-        }
-        subscribe(disposeBag, accountManager.accountDeletedObservable) { [weak self] in
-            self?.handleDeleted(account: $0)
-        }
-        subscribe(disposeBag, accountManager.accountsLostObservable) { [weak self] isAccountsLost in
-            if isAccountsLost {
-                self?.accountsLostRelay.accept(())
-            }
+        subscribe(disposeBag, commonService.sortTypeObservable) { [weak self] _ in
+            self?.handleUpdateSortType()
         }
         subscribe(disposeBag, walletManager.activeWalletsUpdatedObservable) { [weak self] in
             self?.sync(wallets: $0)
@@ -124,19 +74,9 @@ class WalletService {
         _sync(wallets: walletManager.activeWallets)
     }
 
-    private func handleUpdated(account: Account) {
-        if account.id == accountManager.activeAccount?.id {
-            activeAccountRelay.accept(account)
-        }
-    }
-
-    private func handleDeleted(account: Account) {
-        accountRestoreWarningManager.removeIgnoreWarning(account: account)
-    }
-
     private func handleUpdateSortType() {
         queue.async {
-            self.items = self.sorter.sort(balanceItems: self.items, sortType: self.sortType)
+            self.items = self.commonService.sort(balanceItems: self.items)
         }
     }
 
@@ -145,9 +85,9 @@ class WalletService {
     }
 
     private func _sync(wallets: [Wallet]) {
-        let cacheContainer = accountManager.activeAccount.map { cacheManager.cacheContainer(accountId: $0.id) }
+        let cacheContainer = commonService.activeAccount.map { cacheManager.cacheContainer(accountId: $0.id) }
         let priceItemMap = coinPriceService.itemMap(tokens: wallets.map { $0.token })
-        let watchAccount = watchAccount
+        let watchAccount = commonService.watchAccount
 
         let items: [Item] = wallets.map { wallet in
             let item = Item(
@@ -163,7 +103,7 @@ class WalletService {
             return item
         }
 
-        self.items = sorter.sort(balanceItems: items, sortType: sortType)
+        self.items = commonService.sort(balanceItems: items)
         syncTotalItem()
 
         let tokens = Set(wallets.map { $0.token })
@@ -246,7 +186,7 @@ extension WalletService: IWalletAdapterServiceDelegate {
                 balanceDataMap[item.wallet] = balanceData
             }
 
-            self.items = self.sorter.sort(balanceItems: self.items, sortType: self.sortType)
+            self.items = self.commonService.sort(balanceItems: self.items)
             self.syncTotalItem()
 
             self.cacheManager.set(balanceDataMap: balanceDataMap)
@@ -273,8 +213,8 @@ extension WalletService: IWalletAdapterServiceDelegate {
 
             item.balanceData = balanceData
 
-            if self.sortType == .balance, self.items.allSatisfy({ $0.state.isSynced }) {
-                self.items = self.sorter.sort(balanceItems: self.items, sortType: self.sortType)
+            if self.commonService.sortType == .balance, self.items.allSatisfy({ $0.state.isSynced }) {
+                self.items = self.commonService.sort(balanceItems: self.items)
             } else {
                 self.balanceItemUpdatedRelay.accept(item)
             }
@@ -294,8 +234,8 @@ extension WalletService: IWalletAdapterServiceDelegate {
             let oldState = item.state
             item.state = state
 
-            if self.sortType == .balance, self.items.allSatisfy({ $0.state.isSynced }) {
-                self.items = self.sorter.sort(balanceItems: self.items, sortType: self.sortType)
+            if self.commonService.sortType == .balance, self.items.allSatisfy({ $0.state.isSynced }) {
+                self.items = self.commonService.sort(balanceItems: self.items)
             } else {
                 self.balanceItemUpdatedRelay.accept(item)
             }
@@ -304,15 +244,6 @@ extension WalletService: IWalletAdapterServiceDelegate {
                 self.syncTotalItem()
             }
         }
-    }
-
-    func didIgnoreAccountWarning() {
-        guard let account = accountManager.activeAccount, account.nonRecommended else {
-            return
-        }
-
-        accountRestoreWarningManager.setIgnoreWarning(account: account)
-        activeAccountRelay.accept(account)
     }
 
 }
@@ -324,7 +255,7 @@ extension WalletService: IWalletCoinPriceServiceDelegate {
             item.priceItem = priceItemMap[item.wallet.coin.uid]
         }
 
-        items = sorter.sort(balanceItems: items, sortType: sortType)
+        items = commonService.sort(balanceItems: items)
         syncTotalItem()
     }
 
@@ -344,18 +275,6 @@ extension WalletService: IWalletCoinPriceServiceDelegate {
 
 extension WalletService: IWalletService {
 
-    var activeAccountObservable: Observable<Account?> {
-        activeAccountRelay.asObservable()
-    }
-
-    var balanceHiddenObservable: Observable<Bool> {
-        balanceHiddenManager.balanceHiddenObservable
-    }
-
-    var accountsLostObservable: Observable<()> {
-        accountsLostRelay.asObservable()
-    }
-
     var totalItemObservable: Observable<WalletModule.TotalItem?> {
         totalItemRelay.asObservable()
     }
@@ -368,56 +287,12 @@ extension WalletService: IWalletService {
         balanceItemsRelay.asObservable()
     }
 
-    var sortTypeObservable: Observable<WalletModule.SortType> {
-        sortTypeRelay.asObservable()
-    }
-
-    var balancePrimaryValueObservable: Observable<BalancePrimaryValue> {
-        balancePrimaryValueManager.balancePrimaryValueObservable
-    }
-
-    var balancePrimaryValue: BalancePrimaryValue {
-        balancePrimaryValueManager.balancePrimaryValue
-    }
-
-    var balanceHidden: Bool {
-        balanceHiddenManager.balanceHidden
-    }
-
-    var watchAccount: Bool {
-        accountManager.activeAccount?.watchAccount ?? false
-    }
-
-    var activeAccount: Account? {
-        accountManager.activeAccount
-    }
-
-    var isReachable: Bool {
-        reachabilityManager.isReachable
-    }
-
-    var lastCreatedAccount: Account? {
-        accountManager.popLastCreatedAccount()
-    }
-
     func balanceItem(item: WalletModule.Item) -> IBalanceItem? {
         guard let wallet = item.wallet else {
             return nil
         }
 
         return queue.sync { _item(wallet: wallet) }
-    }
-
-    func isCloudBackedUp(account: Account) -> Bool {
-        cloudAccountBackupManager.backedUp(uniqueId: account.type.uniqueId())
-    }
-
-    func notifyAppear() {
-        rateAppManager.onBalancePageAppear()
-    }
-
-    func notifyDisappear() {
-        rateAppManager.onBalancePageDisappear()
     }
 
     func refresh() {
@@ -431,10 +306,6 @@ extension WalletService: IWalletService {
         }
 
         walletManager.delete(wallets: [wallet])
-    }
-
-    func toggleBalanceHidden() {
-        balanceHiddenManager.toggleBalanceHidden()
     }
 
     func toggleConversionCoin() {
